# refactor!: Surfaces: Change Transform3D storage by value