--- a/Core/include/Acts/Geometry/ApproachDescriptor.hpp
+++ b/Core/include/Acts/Geometry/ApproachDescriptor.hpp
@@ -1,14 +1,10 @@
 // This file is part of the Acts project.
 //
-// Copyright (C) 2016-2018 CERN for the benefit of the Acts project
+// Copyright (C) 2016-2020 CERN for the benefit of the Acts project
 //
 // This Source Code Form is subject to the terms of the Mozilla Public
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
-
-///////////////////////////////////////////////////////////////////
-// ApproachDescriptor.h, Acts project
-///////////////////////////////////////////////////////////////////
 
 #pragma once
 
@@ -32,8 +28,6 @@ class BoundaryCheck;
 class ApproachDescriptor {
  public:
   ApproachDescriptor() = default;
-
-  /// Virtual destructor
   virtual ~ApproachDescriptor() = default;
 
   /// @brief Register Layer