--- a/Core/include/Acts/Geometry/CylinderLayer.hpp
+++ b/Core/include/Acts/Geometry/CylinderLayer.hpp
@@ -1,6 +1,6 @@
 // This file is part of the Acts project.
 //
-// Copyright (C) 2016-2018 CERN for the benefit of the Acts project
+// Copyright (C) 2016-2020 CERN for the benefit of the Acts project
 //
 // This Source Code Form is subject to the terms of the Mozilla Public
 // License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -43,7 +43,7 @@ class CylinderLayer : public CylinderSurface, public Layer {
   ///
   /// @return The return object is a shared poiter to the layer.
   static MutableLayerPtr create(
-      const std::shared_ptr<const Transform3D>& transform,
+      const Transform3D& transform,
       const std::shared_ptr<const CylinderBounds>& cbounds,
       std::unique_ptr<SurfaceArray> surfaceArray = nullptr,
       double thickness = 0., std::unique_ptr<ApproachDescriptor> ad = nullptr,
@@ -53,17 +53,10 @@ class CylinderLayer : public CylinderSurface, public Layer {
                                              std::move(ad), laytyp));
   }
 
-  /// Copy constructor - deleted
   CylinderLayer(const CylinderLayer& cla) = delete;
-
-  /// Assignment operator for CylinderLayers - deleted
-  CylinderLayer& operator=(const CylinderLayer&) = delete;
-
-  /// Default Constructor
   CylinderLayer() = delete;
-
-  /// Destructor
   ~CylinderLayer() override = default;
+  CylinderLayer& operator=(const CylinderLayer&) = delete;
 
   /// Transforms the layer into a Surface representation
   /// This is for positioning and extrapolation
@@ -88,7 +81,7 @@ class CylinderLayer : public CylinderSurface, public Layer {
   /// @todo change ApproachDescriptor to unique_ptr
   ///
   /// @return The return object is a shared poiter to the layer.
-  CylinderLayer(const std::shared_ptr<const Transform3D>& transform,
+  CylinderLayer(const Transform3D& transform,
                 const std::shared_ptr<const CylinderBounds>& cBounds,
                 std::unique_ptr<SurfaceArray> surfaceArray = nullptr,
                 double thickness = 0.,