--- a/Core/include/Acts/Geometry/ILayerArrayCreator.hpp
+++ b/Core/include/Acts/Geometry/ILayerArrayCreator.hpp
@@ -7,6 +7,7 @@
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 #pragma once
+
 #include "Acts/Geometry/GeometryContext.hpp"
 #include "Acts/Utilities/BinnedArray.hpp"
 #include "Acts/Utilities/BinningType.hpp"