… Core/include/Acts/Geometry/TrackingVolume.hpp
… Core/include/Acts/Geometry/VolumeBounds.hpp
… Core/include/Acts/Geometry/detail/DefaultDetectorElementBase.hpp
… Core/include/Acts/Geometry/detail/TrackingVolume.ipp
… Core/include/Acts/Surfaces/ConeSurface.hpp
… Core/include/Acts/Surfaces/CylinderSurface.hpp
… Core/include/Acts/Surfaces/DiscSurface.hpp
… Core/include/Acts/Surfaces/LineSurface.hpp
… Core/include/Acts/Surfaces/PerigeeSurface.hpp
… Core/include/Acts/Surfaces/PlaneSurface.hpp
… Core/include/Acts/Surfaces/StrawSurface.hpp
… Core/include/Acts/Surfaces/Surface.hpp
… Core/include/Acts/Surfaces/SurfaceArray.hpp
… Core/include/Acts/Surfaces/detail/Surface.ipp
… Core/include/Acts/Vertexing/ImpactPointEstimator.ipp
… Core/src/Geometry/AbstractVolume.cpp
… Core/src/Geometry/ConeLayer.cpp
… Core/src/Geometry/ConeVolumeBounds.cpp
… Core/src/Geometry/CuboidVolumeBounds.cpp
… Core/src/Geometry/CuboidVolumeBuilder.cpp
… Core/src/Geometry/CutoutCylinderVolumeBounds.cpp
… Core/src/Geometry/CylinderLayer.cpp
… Core/src/Geometry/CylinderVolumeBounds.cpp
… Core/src/Geometry/CylinderVolumeBuilder.cpp
… Core/src/Geometry/CylinderVolumeHelper.cpp
… Core/src/Geometry/DiscLayer.cpp
… Core/src/Geometry/GenericCuboidVolumeBounds.cpp
… Core/src/Geometry/LayerArrayCreator.cpp
… Core/src/Geometry/LayerCreator.cpp
… Core/src/Geometry/PassiveLayerBuilder.cpp