This PR changes the memory model and consistently the construction of Surface/Layer/Volumes to by-value storage and construction from const reference.

Sits on top of #399 (merged)  and #405 (merged).

Fixes #408