@asalzburger 2020-08-31
Comparison of the SurfaceIntersection bench mark:

master (w/o #399 and #405) :

```shell
Running 10 test cases...

Benchmarking theta=0.222428, phi=1.79499...
- Plane: 20000 runs of 2000 iteration(s), 275.6ms total, 13.4400+/-0.0006µs per run, 6.720+/-0.014ns per iteration
- Disc: 20000 runs of 2000 iteration(s), 364.7ms total, 17.2680+/-0.0011µs per run, 8.634+/-0.025ns per iteration
- Cylinder: 20000 runs of 2000 iteration(s), 1605.8ms total, 79.1810+/-0.0060µs per run, 39.590+/-0.133ns per iteration
- Straw: 20000 runs of 2000 iteration(s), 381.7ms total, 17.9860+/-0.0008µs per run, 8.993+/-0.017ns per iteration
```

With (#399, #405 and #411 ):

```shell 
Benchmarking theta=0.222428, phi=1.79499...
- Plane: 20000 runs of 2000 iteration(s), 268.7ms total, 13.2380+/-0.0005µs per run, 6.619+/-0.011ns per iteration
- Disc: 20000 runs of 2000 iteration(s), 334.2ms total, 16.5520+/-0.0004µs per run, 8.276+/-0.008ns per iteration
- Cylinder: 20000 runs of 2000 iteration(s), 1525.3ms total, 75.5540+/-0.0008µs per run, 37.777+/-0.018ns per iteration
- Straw: 20000 runs of 2000 iteration(s), 348.8ms total, 17.1430+/-0.0010µs per run, 8.572+/-0.022ns per iteration
```

Total time went from `105s` to `104s` - so overall looks good.