@asalzburger 2020-08-31
Only the `BinUtility` remains for the moment with a `std::unique_ptr<const Transform3D>` for consistency will be likely changed as well in a subsequent PR.