@asalzburger 2020-09-07
Hi @FabianKlimpel - I have rebased and addressed most of the comments, while I have moved a few others into new issues. Can you have a look if you are ok with these changes, then I can start working off the remaining points in new, smaller PRs.